unify fullcall selection to remove .Type check
previously, `head.html` would check the .Type to load the fullcal styles;
and then various partials would include the matching fullcal scripts to match.
now, `head.html` is parameterized to simultaneously style and scripts.
to approximate the timing of the body loaded scripts, the scripts use "defer" to wait until the page content is ready.